Remove noexcept from function signatures.
This (very slightly) breaks comptibility with the std::atomic interface, but makes the class much more flexible.
I.e. it can inherit from classes that do not put noexcept in their function signatures.

It seems more likely that someone would want to do this than use an ext_atomic in a noexcept class.